Fix test imports to use ../src/ after moving to tests/
